backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
sse-starlette==2.1.3
pydantic==2.9.2
python-dotenv==1.0.1
anthropic
reportlab==4.2.5
